pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-its"
version = "0.1.0"
description = "Agent-based Intelligent Tutoring System with Personalized Memory"
requires-python = ">=3.12"
authors = [
    {name = "Haoran Ye", email = "hrye@stu.pku.edu.cn"},
    {name = "Jiarui Wang", email = "jiarui_wang@seu.edu.cn"},
]
dependencies = [
    "dashscope>=1.24.1",
    "fastapi>=0.104.0",
    "fastmcp>=2.11.2",
    "json-repair>=0.30.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.27",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.2.0",
    "langchain-redis>=0.1.1",
    "langgraph>=0.2.50",
    "langgraph-api>=0.2.3",
    "langgraph-sdk>=0.1.61",
    "langmem>=0.0.28",
    "mem0ai>=0.1.115",
    "pydantic>=2.11.7",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "tavily-python>=0.7.10",
    "uvicorn[standard]>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "chromadb>=1.0.15",
    "jupyter>=1.0.0",
    "pre-commit>=3.5.0",
    "pytest-cov>=4.1.0",
]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true

